+++ src/App.jsx
@@ -5,6 +5,8 @@ import ToDoList from "./pages/ToDoList";
 import Login from "./pages/Login";
 import Signup from "./pages/Signup";
 import useAuth from "./hooks/useAuth";
+import ProductDetail from "./pages/ProductDetail";
+import ProductsList from "./pages/ProductsList";
 
 function App() {
   const { user, handleLogin, handleLogout } = useAuth();
@@ -14,6 +16,10 @@ function App() {
       <Header user={user} handleLogout={handleLogout} />
       <Routes>
         <Route path="/" element={<HomePage />} />
+        <Route path="products">
+          <Route index element={<ProductsList />} />
+          <Route path=":productSlug" element={<ProductDetail />} />
+        </Route>
         <Route path="/todolist" element={<ToDoList />} />
         <Route path="/login" element={<Login onLogin={handleLogin} />} />
         <Route path="/signup" element={<Signup />} />
